breaking into smaller components
