+++ src/views/Home/index.vue
@@ -4,88 +4,19 @@
       <h1 class="title">-SmartSignature-</h1>
       <h2 class="subtitle">首个EOS去中心化智能签名项目</h2>
     </div>
-    <div class="my-banner">
-      <div class="my-stat">
-        <div class="logined" v-if="isLogined">
-          <p
-            @click="$router.push({ name: 'User', params: {username: currentUsername } })"
-            class="username"
-          >{{currentUsername}}</p>
-          <p class="my-balance">
-            {{eosBalance}}
-            <span class="coin-symbol">EOS</span>
-          </p>
-        </div>
-        <div class="not-login-yet" v-else>
-          <p>欢迎来到 Smart Signature</p>
-          <za-button block theme="primary" @click="$router.push({name: 'Login'})">使用钱包登录</za-button>
-        </div>
-      </div>
-      <za-button block bordered 
-       @click="$router.push({name: 'About'})">玩法介绍</za-button>
-    </div>
-    <div class="articles">
-      <za-tabs v-model="activeNameSwipe" @change="handleClick">
-        <za-tab-pane :label="tab.label" :name="tab.label" v-for="tab in tabs" :key="tab.label">
-          <div class="content">
-            <ArticleCard :article="a" v-for="a in articles" :key="a.id" />
-          </div>
-        </za-tab-pane>
-      </za-tabs>
-    </div>
+    <MyBanner />
+    <ArticlesList/>
   </div>
 </template>
 
 <script>
-import axios from 'axios';
-import { ArticleCard } from '@/components/';
-import { mapState, mapActions, mapGetters } from 'vuex';
+import MyBanner from "./MyBanner.vue";
+import ArticlesList from "./ArticlesList.vue";
 
 export default {
-  name: 'home',
-  computed: {
-    ...mapState(['scatterAccount', 'balances']),
-    ...mapGetters(['currentUsername']),
-    eosBalance() {
-      return this.balances.eos.slice(0, -4);
-    },
-    isLogined() {
-      return this.scatterAccount !== null;
-    },
-  },
-  components: { ArticleCard },
-  created() {
-    this.getArticlesList();
-  },
-  methods: {
-    ...mapActions(['loginScatterAsync']),
-    async getArticlesList() {
-      const articles = 'https://smartsignature.azurewebsites.net/api/article';
-      const { data } = await axios.get(articles);
-      this.articles = data;
-    },
-    handleClick(tab, event) {
-      console.log(tab, event);
-    },
-  },
-  data() {
-    return {
-      articles: [],
-      activeNameSwipe: '文章列表',
-      selectedLabelDefault: '文章列表',
-      tabs: [
-        {
-          label: '文章列表',
-        },
-        {
-          label: '最多支持',
-        },
-        {
-          label: '最多分享',
-        },
-      ],
-    };
-  },
+  name: "Home",  
+  components: { ArticlesList, MyBanner },
+  data: () => ({})
 };
 </script>
 
@@ -109,9 +40,6 @@ h1.title {
   line-height: 27px;
   letter-spacing: 2px;
 }
-.username {
-  font-weight: bold;
-}
 h2.subtitle {
   font-size: 16px;
   font-family: PingFangSC-Light;
@@ -119,44 +47,5 @@ h2.subtitle {
   color: rgba(255, 255, 255, 1);
   line-height: 18px;
   letter-spacing: 1px;
-}
-.my-banner {
-  margin: auto;
-  margin-top: -32px;
-  text-align: center;
-  max-width: 335px;
-  /* margin: -32px 20px 0 20px; */
-  height: 220px;
-  padding: 8px;
-  background: rgba(255, 255, 255, 1);
-  box-shadow: 0px 5px 5px 0px rgba(213, 213, 213, 0.5);
-  border-radius: 4px;
-}
-.my-stat {
-  text-align: left;
-  /* float: left; */
-}
-.coin-symbol {
-  color: #999999;
-}
-.my-balance {
-  font-size: 16px;
-  font-family: PingFang SC, STHeitiSC-Light, Helvetica-Light, arial, sans-serif;
-  font-weight: 600;
-  color: rgba(0, 0, 0, 1);
-  line-height: 22px;
-  letter-spacing: 1px;
-}
-
-.articles {
-  /* background: rgba(240, 240, 240, 1); */
-  margin-top: -100px;
-  padding-top: 100px;
-}
-.article {
-  text-align: left;
-}
-.card {
-  margin: 5px;
 }
 </style>
